+++ requirements.txt
@@ -7,9 +7,8 @@ click==8.1.7
 colorama==0.4.6
 dnspython==2.6.1
 email-validator==2.1.0.post1
-environs==11.0.0
 fastapi==0.109.2
-fastapi_cors==0.0.6
+fastapi-limiter==0.1.6
 greenlet==3.0.3
 h11==0.14.0
 httpcore==1.0.3
@@ -20,22 +19,21 @@ iniconfig==2.0.0
 itsdangerous==2.1.2
 Jinja2==3.1.3
 MarkupSafe==2.1.5
-marshmallow==3.21.0
 orjson==3.9.14
-packaging==23.2
-pluggy==1.4.0
+packaging==24.0
+pluggy==1.5.0
 psycopg2==2.9.9
 pydantic==2.6.1
 pydantic-extra-types==2.5.0
 pydantic-settings==2.2.1
 pydantic_core==2.16.2
 pypng==0.20220715.0
-pytest==8.0.2
-pytest-mock==3.14.0
+pytest==8.2.2
 python-dateutil==2.8.2
 python-dotenv==1.0.1
 python-multipart==0.0.9
 PyYAML==6.0.1
+redis==5.1.0b6
 segno==1.6.1
 six==1.16.0
 sniffio==1.3.0
